requirements.txt
pillow==11.0.0
scikit-learn==1.5.2
tensorflow==2.12.0
scikit-image==0.24.0
tqdm==4.66.5
matplotlib==3.9.2